refactor: Update ClearShaderCache with confirmation dialogs and remove utils dependency

+++ Packages/io.github.ykysnk.yky-toolkit/Editor/ClearShaderCache.cs
@@ -1,5 +1,4 @@
 using System.IO;
-using io.github.ykysnk.utils;
 using UnityEditor;
 using UnityEngine;
 
@@ -10,8 +9,11 @@ internal static class ClearShaderCache
     [MenuItem($"Tools/{Util.Name}/Clear Shader Cache")]
     private static void Clear()
     {
+        if (!EditorUtility.DisplayDialog("Clear Shader Cache", "Are you sure you want to clear the Shader Cache?", "Yes",
+                "No")) return;
         var shaderCachePath = Path.Combine(Application.dataPath, "../Library/ShaderCache");
         if (Directory.Exists(shaderCachePath)) Directory.Delete(shaderCachePath, true);
-        Utils.Log(nameof(ClearShaderCache), $"Clearing Shader Cache at {shaderCachePath}");
+        EditorUtility.DisplayDialog("Shader Cache Cleared", "Shader Cache has been cleared.", "OK");
+        AssetDatabase.Refresh();
     }
 }
